refactor: 차 이름 CarName 클래스로 분리
